Fix values clearing when you go to the account panel when logged out

+++ Assets/Scripts/AccountMenu/AccountLoggedOut.cs
@@ -7,6 +7,7 @@ public class AccountLoggedOut : MonoBehaviour
     public Button loggedOutLoginButton;
     public Button loggedOutRegisterButton;
     public Button loggedOutBackButton;
+    public bool clearValues = false;
 
     void Awake()
     {
@@ -17,19 +18,23 @@ public class AccountLoggedOut : MonoBehaviour
 
     void OnEnable()
     {
-        PlayerPrefs.DeleteKey("gameSession");
-        PlayerPrefs.DeleteKey("userName");
-        PlayerPrefs.DeleteKey("userId");
-        PlayerPrefs.SetString("HighScoreV2", "0");
-        PlayerPrefs.SetInt("icon", 1);
-        PlayerPrefs.SetInt("overlay", 0);
-        PlayerPrefs.DeleteKey("TotalNormalBerries");
-        PlayerPrefs.DeleteKey("TotalPoisonBerries");
-        PlayerPrefs.DeleteKey("TotalSlowBerries");
-        PlayerPrefs.DeleteKey("TotalUltraBerries");
-        PlayerPrefs.DeleteKey("TotalSpeedyBerries");
-        PlayerPrefs.DeleteKey("TotalAttempts");
-        PlayerPrefs.DeleteKey("BirdColor");
-        PlayerPrefs.DeleteKey("OverlayColor");
+        if (clearValues)
+        {
+            PlayerPrefs.DeleteKey("gameSession");
+            PlayerPrefs.DeleteKey("userName");
+            PlayerPrefs.DeleteKey("userId");
+            PlayerPrefs.SetString("HighScoreV2", "0");
+            PlayerPrefs.SetInt("icon", 1);
+            PlayerPrefs.SetInt("overlay", 0);
+            PlayerPrefs.DeleteKey("TotalNormalBerries");
+            PlayerPrefs.DeleteKey("TotalPoisonBerries");
+            PlayerPrefs.DeleteKey("TotalSlowBerries");
+            PlayerPrefs.DeleteKey("TotalUltraBerries");
+            PlayerPrefs.DeleteKey("TotalSpeedyBerries");
+            PlayerPrefs.DeleteKey("TotalAttempts");
+            PlayerPrefs.DeleteKey("BirdColor");
+            PlayerPrefs.DeleteKey("OverlayColor");
+            clearValues = false;
+        }
     }
 }
